requirements.txt
customtkinter
youtube-transcript-api
google-genai
python-dotenv
